requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
jupyter>=1.0.0
plotly>=5.0.0
notebook>=6.4.0
ipykernel>=6.0.0
sklearn-pandas>=2.0.0
